Use LibUSB-1.0 RawHID backend for Linux
Replaces LibUSB-0.1 backend.
